Challenge Detail View 기본 UI 완성
