+++ lib/src/app/view/app.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:yum_application/src/app/viewModel/app_view_model.dart';
-import 'package:yum_application/src/challege/view/challenge_list_view.dart';
+import 'package:yum_application/src/challenge/view/challenge_list_view.dart';
 import 'package:yum_application/src/common/image_widget.dart';
 import 'package:yum_application/src/ingredient/view/home_view.dart';
