+++ lib/src/challenge/view/challenge_case_view.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import 'package:yum_application/src/challege/widget/challenge_detail_view.dart';
+import 'package:yum_application/src/challenge/widget/challenge_detail_view.dart';
 import 'package:yum_application/src/common/image_widget.dart';
 import 'package:yum_application/src/ingredient/viewModel/ingredient_view_model.dart';
-import 'package:yum_application/src/challege/widget/challenge_row.dart';
+import 'package:yum_application/src/challenge/widget/challenge_row.dart';
 import 'package:yum_application/src/ingredient/widget/single_button.dart';
 
 class ChallengeCaseView extends StatelessWidget {
