+++ lib/src/challenge/widget/challenge_detail_view.dart
@@ -15,38 +15,16 @@ class ChallengeDetailView extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      backgroundColor: Theme.of(context).colorScheme.tertiary,
-      appBar: AppBar(
-        backgroundColor: Theme.of(context).colorScheme.onPrimaryContainer,
-        foregroundColor: Theme.of(context).colorScheme.onSecondary,
-        leading: Builder(
-          builder: (context) {
-            return GestureDetector(
-              onTap: () {
-                Navigator.of(context).pop();
-              },
-              child: const Icon(
-                Icons.arrow_back_ios,
-              ),
-            );
-          },
-        ),
-        elevation: 0.0,
-        title: Text(
-          "챌린지",
-          style: Theme.of(context).textTheme.headlineLarge,
-        ),
-      ),
-      body: SingleChildScrollView(
-        child: Column(
-          children: [
+      backgroundColor: Colors.white,
+      appBar: _appBar(),
+      body: Container(
+        color: Theme.of(context).colorScheme.tertiary,
+        child: CustomScrollView(
+          slivers: [
             _header(imagePath),
             _body(),
             _bottom(),
-            SingleButton(
-              text: "레시피 등록하러 가기",
-              onTap: () {},
-            )
+            _button(),
           ],
         ),
       ),
@@ -54,39 +32,71 @@ class ChallengeDetailView extends StatelessWidget {
   }
 }
 
+PreferredSizeWidget _appBar() {
+  return PreferredSize(
+    preferredSize: AppBar().preferredSize,
+    child: Builder(builder: (context) {
+      return AppBar(
+        backgroundColor: Theme.of(context).colorScheme.onPrimaryContainer,
+        foregroundColor: Theme.of(context).colorScheme.onSecondary,
+        elevation: 0.0,
+        title: Text(
+          "챌린지 상세",
+          style: Theme.of(context).textTheme.headlineLarge,
+        ),
+      );
+    }),
+  );
+}
+
 Widget _header(String imagePath) => Builder(builder: (context) {
-      return Container(
-        width: double.infinity,
-        decoration: BoxDecoration(
-            color: Theme.of(context).colorScheme.onPrimaryContainer,
-            borderRadius: const BorderRadius.only(
-              bottomLeft: Radius.circular(20),
-              bottomRight: Radius.circular(20),
-            )),
-        child: Column(
+      return SliverAppBar(
+        toolbarHeight: 220,
+        floating: true,
+        pinned: false,
+        backgroundColor: Colors.white,
+        shape: const RoundedRectangleBorder(
+          borderRadius: BorderRadius.vertical(bottom: Radius.circular(20)),
+        ),
+        flexibleSpace: Column(
           children: [
-            SizedBox(
+            Container(
               width: double.infinity,
-              height: 138,
-              child: FittedBox(
-                  fit: BoxFit.contain, child: ImageWidget(path: imagePath)),
-            ),
-            SizedBox(
-              width: double.infinity,
-              child: Padding(
-                padding: const EdgeInsets.only(
-                    top: 27, left: 20, right: 20, bottom: 20),
-                child: Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    Text(
-                      "레시피 사용 횟수 도전!",
-                      style: Theme.of(context).textTheme.headlineLarge,
+              decoration: BoxDecoration(
+                color: Theme.of(context).colorScheme.onPrimaryContainer,
+              ),
+              child: Column(
+                children: [
+                  SizedBox(
+                    height: 138,
+                    child: FittedBox(
+                      fit: BoxFit.contain,
+                      child: ImageWidget(path: imagePath),
                     ),
-                    Text("레시피 사용 횟수 챌린지의 히든 요리는 무엇일까요?",
-                        style: Theme.of(context).textTheme.bodySmall),
-                  ],
-                ),
+                  ),
+                  Padding(
+                    padding: const EdgeInsets.only(left: 20, bottom: 18),
+                    child: SizedBox(
+                      width: double.infinity,
+                      child: Column(
+                        crossAxisAlignment: CrossAxisAlignment.start,
+                        children: [
+                          Padding(
+                            padding: const EdgeInsets.only(bottom: 4),
+                            child: Text(
+                              "레시피 사용 횟수 도전!",
+                              style: Theme.of(context).textTheme.headlineLarge,
+                            ),
+                          ),
+                          Text(
+                            "레시피 사용 횟수 챌린지의 히든 요리는 무엇일까요?",
+                            style: Theme.of(context).textTheme.labelSmall,
+                          ),
+                        ],
+                      ),
+                    ),
+                  ),
+                ],
               ),
             ),
           ],
@@ -94,12 +104,13 @@ Widget _header(String imagePath) => Builder(builder: (context) {
       );
     });
 
-Widget _body() => Padding(
+Widget _body() {
+  return SliverToBoxAdapter(
+    child: Padding(
       padding: const EdgeInsets.symmetric(horizontal: 20, vertical: 20),
       child: Builder(builder: (context) {
         return Container(
           width: double.infinity,
-          height: 171,
           decoration: BoxDecoration(
             color: Theme.of(context).colorScheme.onPrimaryContainer,
             borderRadius: BorderRadius.circular(20),
@@ -109,20 +120,56 @@ Widget _body() => Padding(
             children: [
               Text(
                 "3단계 도전중",
-                style: Theme.of(context).textTheme.titleSmall,
+                style: Theme.of(context).textTheme.headlineSmall,
+              ),
+              Text(
+                "3개 달성! 다음 단계까지 2개 남았어요",
+                style: Theme.of(context).textTheme.labelSmall,
+              ),
+              Padding(
+                padding: const EdgeInsets.only(
+                    left: 16, right: 16, top: 20, bottom: 41),
+                child: Stack(
+                  children: [
+                    Container(
+                      width: double.infinity,
+                      height: 20,
+                      decoration: BoxDecoration(
+                        color: Theme.of(context).colorScheme.secondaryContainer,
+                        borderRadius: BorderRadius.circular(20),
+                      ),
+                    ),
+                    Positioned(
+                      left: 0,
+                      top: 0,
+                      bottom: 0,
+                      child: CircleAvatar(
+                        radius: 10,
+                        backgroundColor:
+                            Theme.of(context).colorScheme.secondary,
+                      ),
+                    ),
+                  ],
+                ),
               )
             ],
           ),
         );
       }),
-    );
+    ),
+  );
+}
 
-Widget _bottom() => Padding(
-      padding: const EdgeInsets.only(left: 20, right: 20, bottom: 32),
+Widget _bottom() {
+  return SliverToBoxAdapter(
+    child: Padding(
+      padding: const EdgeInsets.only(
+        left: 20,
+        right: 20,
+      ),
       child: Builder(builder: (context) {
         return Container(
           width: double.infinity,
-          height: 171,
           decoration: BoxDecoration(
             color: Theme.of(context).colorScheme.onPrimaryContainer,
             borderRadius: BorderRadius.circular(20),
@@ -132,10 +179,55 @@ Widget _bottom() => Padding(
             children: [
               Text(
                 "레시피 리뷰 작성 발자취",
-                style: Theme.of(context).textTheme.titleSmall,
-              )
+                style: Theme.of(context).textTheme.headlineSmall,
+              ),
+              const SizedBox(height: 12),
+              GridView.builder(
+                shrinkWrap: true,
+                physics: const NeverScrollableScrollPhysics(),
+                gridDelegate: const SliverGridDelegateWithFixedCrossAxisCount(
+                  crossAxisCount: 4,
+                  crossAxisSpacing: 8,
+                  mainAxisSpacing: 8,
+                  childAspectRatio: 1,
+                ),
+                itemCount: 12,
+                itemBuilder: (context, index) {
+                  return ImageWidget(path: "assets/sample_$index.png");
+                },
+              ),
             ],
           ),
         );
       }),
-    );
+    ),
+  );
+}
+
+Widget _button() {
+  return SliverToBoxAdapter(
+    child: Builder(builder: (context) {
+      return Column(
+        children: [
+          SizedBox(
+            width: double.infinity,
+            child: Padding(
+              padding: const EdgeInsets.symmetric(vertical: 10, horizontal: 20),
+              child: Text(
+                "챌린지를 완료하고 싶나요?",
+                style: Theme.of(context).textTheme.labelSmall,
+              ),
+            ),
+          ),
+          Padding(
+            padding: const EdgeInsets.only(bottom: 20),
+            child: SingleButton(
+              text: "레시피 등록하러 가기",
+              onTap: () {},
+            ),
+          ),
+        ],
+      );
+    }),
+  );
+}
