+++ test/challenge/view/challenge_list_view_test.dart
@@ -3,7 +3,7 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/annotations.dart';
 import 'package:provider/provider.dart';
 import 'package:yum_application/src/data/ingredient/repository/ingredient_repository.dart';
-import 'package:yum_application/src/challege/view/challenge_list_view.dart';
+import 'package:yum_application/src/challenge/view/challenge_list_view.dart';
 import 'package:yum_application/src/ingredient/viewModel/ingredient_view_model.dart';
 import '../../common/mock_navigator_observer.dart';
 import '../../ingredient/view/home_view_test.mocks.dart';
